+++ cli/bin/modus.js
@@ -1,4 +1,4 @@
-#!/usr/bin/env node --no-warnings=ExperimentalWarning
+#!/usr/bin/env node
 import { execute } from "@oclif/core";
 
 await execute({ dir: import.meta.url });