+++ cli/src/commands/sdk/remove/index.ts
@@ -7,7 +7,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import { Args, Command, Flags } from "@oclif/core";
+import { Args, Flags } from "@oclif/core";
 import chalk from "chalk";
 
 import * as fs from "../../../util/fs.js";
@@ -15,12 +15,14 @@ import * as vi from "../../../util/versioninfo.js";
 import { parseSDK, SDK } from "../../../custom/globals.js";
 import { withSpinner } from "../../../util/index.js";
 import * as inquirer from "@inquirer/prompts";
+import { BaseCommand } from "../../../baseCommand.js";
+import { isErrorWithName } from "../../../util/errors.js";
 
-export default class SDKRemoveCommand extends Command {
+export default class SDKRemoveCommand extends BaseCommand {
   static args = {
     name: Args.string({
-      description: "SDK name to remove (or 'all' to remove all SDKs)",
-      required: true,
+      description: "SDK name to remove",
+      options: ["all", "go", "golang", "assemblyscript", "as"],
     }),
     version: Args.string({
       description: "SDK version to remove, if removing a specific SDK. Leave blank to remove all versions of the SDK.",
@@ -29,11 +31,6 @@ export default class SDKRemoveCommand extends Command {
   };
 
   static flags = {
-    help: Flags.help({
-      char: "h",
-      helpLabel: "-h, --help",
-      description: "Show help message",
-    }),
     runtimes: Flags.boolean({
       char: "r",
       default: false,
@@ -52,6 +49,58 @@ export default class SDKRemoveCommand extends Command {
   async run(): Promise<void> {
     try {
       const { args, flags } = await this.parse(SDKRemoveCommand);
+
+      if (!args.name) {
+        const goSdkVersions = await vi.getInstalledSdkVersions(SDK.Go);
+        const asSdkVersions = await vi.getInstalledSdkVersions(SDK.AssemblyScript);
+
+        if (goSdkVersions.length === 0 && asSdkVersions.length === 0) {
+          this.log(chalk.yellow("No Modus SDKs are installed."));
+          this.exit(1);
+        }
+
+        const sdkVersions: { sdk: SDK; version: string }[] = [...goSdkVersions.map((version) => ({ sdk: SDK.Go, version })), ...asSdkVersions.map((version) => ({ sdk: SDK.AssemblyScript, version }))];
+        if (goSdkVersions.length > 0) {
+          sdkVersions.push({
+            sdk: SDK.Go,
+            version: "all",
+          });
+        }
+        if (asSdkVersions.length > 0) {
+          sdkVersions.push({
+            sdk: SDK.AssemblyScript,
+            version: "all",
+          });
+        }
+
+        const sdkVersionToRemove = await inquirer.select({
+          message: "Select a SDK version to remove",
+          choices: sdkVersions.map((item) => ({
+            name: item.version === "all" ? chalk.bold(`All ${item.sdk} SDK versions`) : `${item.sdk} SDK ${item.version}`,
+            value: item,
+          })),
+        });
+
+        if (sdkVersionToRemove.version === "all") {
+          const versions = await vi.getInstalledSdkVersions(sdkVersionToRemove.sdk);
+          const confirmed = await inquirer.confirm({
+            message: `Are you sure you want to remove all Modus ${sdkVersionToRemove.sdk} SDKs?`,
+            default: false,
+          });
+          if (!confirmed) {
+            this.abort();
+          }
+
+          for (const version of versions) {
+            await this.removeSDK(sdkVersionToRemove.sdk, version);
+          }
+        } else {
+          await this.removeSDK(parseSDK(sdkVersionToRemove.sdk), sdkVersionToRemove.version);
+        }
+
+        return;
+      }
+
       if (!args.version) {
         this.logError(`No SDK specified! Run ${chalk.whiteBright("modus sdk remove <name> [version]")}, or ${chalk.whiteBright("modus sdk remove all")}`);
         return;
@@ -136,8 +185,8 @@ export default class SDKRemoveCommand extends Command {
           await this.removeSDK(sdk, args.version);
         }
       }
-    } catch (err: any) {
-      if (err.name === "ExitPromptError") {
+    } catch (err) {
+      if (isErrorWithName(err) && err.name === "ExitPromptError") {
         this.abort();
       } else {
         throw err;