+++ cli/src/util/errors.ts
@@ -1,0 +1,12 @@
+/*
+ * Copyright 2024 Hypermode Inc.
+ * Licensed under the terms of the Apache License, Version 2.0
+ * See the LICENSE file that accompanied this code for further details.
+ *
+ * SPDX-FileCopyrightText: 2024 Hypermode Inc. <hello@hypermode.com>
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export function isErrorWithName(err: unknown): err is { name: string } {
+  return typeof err === "object" && err !== null && "name" in err;
+}