+++ cli/src/util/systemVersions.ts
@@ -7,12 +7,11 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
+import process from "node:process";
 import { execFile } from "./cp.js";
-import { ModusHomeDir } from "../custom/globals.js";
 
 const EXEC_OPTIONS = {
   shell: true,
-  cwd: ModusHomeDir,
   env: process.env,
 };
 
@@ -22,17 +21,28 @@ export async function getGoVersion(): Promise<string | undefined> {
     const parts = result.stdout.split(" ");
     const str = parts.length > 2 ? parts[2] : undefined;
     if (str?.startsWith("go")) {
-      return str.slice(2);
+      const ver = str.slice(2);
+
+      // if version is two parts, add a .0 to make it semver compatible
+      return ver.split(".").length === 2 ? ver + ".0" : ver;
     }
-  } catch {}
+  } catch {
+    /* empty */
+  }
 }
 
 export async function getTinyGoVersion(): Promise<string | undefined> {
   try {
     const result = await execFile("tinygo", ["version"], EXEC_OPTIONS);
     const parts = result.stdout.split(" ");
-    return parts.length > 2 ? parts[2] : undefined;
-  } catch {}
+    const ver = parts.length > 2 ? parts[2] : undefined;
+    if (!ver) return undefined;
+
+    // if version is two parts, add a .0 to make it semver compatible
+    return ver.split(".").length === 2 ? ver + ".0" : ver;
+  } catch {
+    /* empty */
+  }
 }
 
 export async function getNPMVersion(): Promise<string | undefined> {
@@ -40,5 +50,7 @@ export async function getNPMVersion(): Promise<string | undefined> {
     const result = await execFile("npm", ["--version"], EXEC_OPTIONS);
     const parts = result.stdout.split(" ");
     return parts.length > 0 ? parts[0].trim() : undefined;
-  } catch {}
+  } catch {
+    /* empty */
+  }
 }