+++ runtime/main.go
@@ -10,6 +10,8 @@
 package main
 
 import (
+	"context"
+
 	"github.com/hypermodeinc/modus/runtime/app"
 	"github.com/hypermodeinc/modus/runtime/config"
 	"github.com/hypermodeinc/modus/runtime/envfiles"
@@ -24,6 +26,9 @@ func main() {
 	// Initialize the configuration
 	config.Initialize()
 
+	// Create the main background context
+	ctx := context.Background()
+
 	// Initialize the logger
 	log := logger.Initialize()
 	log.Info().
@@ -31,8 +36,10 @@ func main() {
 		Str("environment", config.GetEnvironmentName()).
 		Msg("Starting Modus Runtime.")
 
-	// Load environment variables from .env file(s)
-	envfiles.LoadEnvFiles(log)
+	err := envfiles.LoadEnvFiles(ctx)
+	if err != nil {
+		log.Warn().Err(err).Msg("Failed to load environment files.")
+	}
 
 	// Initialize Sentry (if enabled)
 	rootSourcePath := app.GetRootSourcePath()
@@ -40,7 +47,7 @@ func main() {
 	defer utils.FlushSentryEvents()
 
 	// Start the background services
-	ctx := services.Start()
+	ctx = services.Start(ctx)
 	defer services.Stop(ctx)
 
 	// Set local mode in development