+++ runtime/services/services.go
@@ -31,9 +31,7 @@ import (
 )
 
 // Starts any services that need to be started when the runtime starts.
-func Start() context.Context {
-	// Create the main background context
-	ctx := context.Background()
+func Start(ctx context.Context) context.Context {
 
 	// Note, we cannot start a Sentry transaction here, or it will also be used for the background services, post-initiation.